split code into components
