+++ app/components/class-collection.hbs
@@ -1,0 +1,21 @@
+<div class="au-u-flex au-u-flex--column au-u-flex--center au-u-1-1">
+  {{#each @document as |collection|}}
+  {{! First level: collections }}
+    <CustomAccordion @defaultOpen="true">
+      <:toolbar>
+        {{collection.className}}
+        <p class="au-u-para-tiny">
+          <AuLinkExternal @skin="primary" href={{@property.path}} class="truncate-url" title={{@property.path}}>
+            {{collection.classURI}}
+          </AuLinkExternal>
+        </p>
+      </:toolbar>
+      <:counts>{{collection.count}}</:counts>
+      <:content>
+        {{#each collection.objects as |object index|}}
+          <RecursiveEntry @subject={{object}} @index={{index}}/>
+        {{/each}}
+      </:content>
+    </CustomAccordion>
+  {{/each}}
+</div>
