+++ app/components/subject-property.hbs
@@ -1,0 +1,23 @@
+<CustomAccordion @defaultOpen="true">
+  <:toolbar>
+    <p class="au-u-h4">{{@property.name}}</p>
+    <p class="au-u-para-tiny">
+      <AuLinkExternal @skin="primary" href={{@property.path}} class="truncate-url" title={{@property.path}}>
+        {{@property.path}}
+      </AuLinkExternal>
+    </p>
+    <AuAlert @skin="default">
+      <p class="au-u-para">{{@property.description}}</p>
+    </AuAlert>
+  </:toolbar>
+  <:counts>{{this.actualCount}}</:counts>
+  <:content>
+    {{#each @property.value as |value|}}
+      {{#if value.uri}}
+        <RecursiveEntry @subject={{value}} />
+      {{else}}
+        <p>{{this.displayValue}}</p>
+      {{/if}}
+    {{/each}}
+  </:content>
+</CustomAccordion>
